+++ shared/src/commonMain/kotlin/ai/lerna/multiplatform/config/ReadWriteFile.kt
@@ -3,6 +3,6 @@ package ai.lerna.multiplatform.config
 import org.jetbrains.kotlinx.multik.ndarray.data.D2Array
 
 expect class ReadWriteFile constructor(){
-    fun read(path : String, filename: String) : List<DoubleArray>
+    fun read(path : String, filename: String) : List<FloatArray>
     fun save(path : String, filename : String, nbLine : Int)
 }
